backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
cohere>=5.5.3
qdrant-client>=1.9.1
python-dotenv>=1.0.0
pydantic>=2.8.2
httpx>=0.25.2
cachetools>=5.3.2
fastembed